doc: Adjust definition store plain text file format
Experimenting with Pest grammar I realized that allowing some values to
have any character, except new lines, were very hard, if not impossible,
or complicated the grammar to certain substantial degree.

For avoiding the complexity and retain the validation of the values as
much as possible in the same grammar for not having to implement those
many validation rules in code, I redefined the format of certain fields'
values to split them in lines rather than using a specific character.

I redefined the `notes` field to require to indent the lines belonging
to its value to be clear for the reader.

The `permissions` field format is redesigned to support specifying
permissions at the project's root level because Storj access grants
allows it and I missed it in the initial definition.

I added my reasons and the intention why the file format is very strict.